requirements.txt
requests>=2.31.0
mwparserfromhell>=0.6.4
python-dotenv>=1.0.0